Update brainfuck.class.php
Since php can traverse strings in the same manner as arrays, it's pointless to convert the code to an array.

As well, there is no point in deleting non-parseable characters, as they can simply be ignored. This won't have a significant performance impact unless there is a large amount of garbage in the file.